Add BM::Instrumentations::Rack
