+++ lib/bm/instrumentations/rack/endpoint.rb
@@ -2,7 +2,7 @@
 
 module BM
   module Instrumentations
-    module Rack
+    class Rack
       # Is a Rack env key for a name of an endpoint which served a request
       ENDPOINT = 'x.rack.endpoint'
     end
